requirements.txt
pytorch-lightning==1.5.1
hydra-core==1.1.0
hydra_colorlog
wandb
rich
